added fonts.o, bypassing dat2c/ i386 dependence
So fonts.dat was actually just a file that held several fonts the game uses. dat2s was called in the configure and make file when fonts.o was not present, translating fonts.dat into assembly. I spent the better part of my day checking the assembly code, and it just looks like they were trying to save space for sourceforge to me. 

After the dat2s conversion, it took the resultant fonts.s and ran it through gcc to make it into an object file usable by the game.